requirements.txt
Flask
gunicorn
scikit-learn
pandas
joblib
numpy
PyPDF2
python-docx